+++ src/components/CustomTable/Cells/EditButtonCell/EditButtonCell.js
@@ -1,4 +1,4 @@
-import EditButton from "@/components/EditButton/EditButton"
+import EditButton from "@/components/Buttons/EditButton"
 import { TableCell, withStyles } from "@material-ui/core"
 
 const NewTableCell = withStyles((theme) => ({
